refactor: Update import statements in collection module controllers
This commit updates the import statements in the collection module's controllers to import the AuthType enum from the correct location. This ensures that the code is using the updated AuthType enum and avoids any potential import errors. The changes improve the maintainability and readability of the code.
