+++ src/modules/collection/interface/collection-team.controller.ts
@@ -11,11 +11,11 @@ import {
 
 import { IPromiseResponse } from '@/common/response_service/interface/response.interface';
 import { Auth } from '@/modules/auth/application/decorator/auth.decorator';
+import { AuthType } from '@/modules/auth/domain/auth_type.enum';
 import { CreateEnvironmentsDto } from '@/modules/enviroment/application/dto/create-all-environments.dto';
 import { EnviromentResponseDto } from '@/modules/enviroment/application/dto/enviroment-response.dto';
 import { Enviroment } from '@/modules/enviroment/domain/enviroment.domain';
 import { FolderResponseDto } from '@/modules/folder/application/dto/folder-response.dto';
-import { AuthType } from '@/modules/user/domain/auth_type.enum';
 
 import { CollectionResponseDto } from '../application/dto/collection-response.dto';
 import { CreateCollectionDto } from '../application/dto/create-collection.dto';
